refactor: remove duplicate oauthprovider and import from config/entities (#719)
