+++ src/google/config.ts
@@ -1,4 +1,4 @@
-import { OAuthProvider } from "../auth/types/provider";
+import { OAuthProvider } from "../config/entities";
 import { GoogleIcon } from "../components/common/CustomIcon/components/GoogleIcon/googleIcon";
 import { GOOGLE_SIGN_IN_PROVIDER_ID } from "./constants";
 import { GoogleProfile } from "./types";
